platform: stm32mp2: Fix compilation warning on array subscript of stack seal region
The stack seal region is defined as an 8 byte region in the linker
scripts, then declared as an extern uint32_t i.e 4 bytes value.

The compiler generates a warning when this is cast to a uint64_t *
in order to write the 8 byte stack seal value.

Correctly declare the stack seal region as an 8 byte region.

The CMSIS stack seal function accepts an uint32_t pointer then
itself cast it back to a uint64_t pointer.

Change-Id: I4329b97b6e93ece2c45ba5047cb3a727f9dd4b53
Signed-off-by: Ludovic Barre <ludovic.barre@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/363994