platform: stm32mp2: bsec3: manage upper OTP keys with dummy provisioning
Upper OTP fuses cannot be accessed if the chip is not in Secure Lock
state. In this case, use the mirror_dummy struct for the dummy IAK
and entropy seed. This only has to be done when TFM_DUMMY_PROVISIONING
flag is set and not in BL2 (which uses the BSEC mirror).

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: Ibbf333e8c746d93aaadc7297fceaa43ae7b6f6f3
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/427415
Reviewed-by: Ludovic BARRE <ludovic.barre@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>