platform: stm32mp2: bsec3: correct stm32_bsec_shadow_init
There was a misalignment between the values passed to _otp_write(),
the later one needs to be multiplied by 4 (sizeof(uint32_t)). But
it wasn't seen as this was already written by another function, and the
error is not caught there. Correct the call and add a check on return
value.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Signed-off-by: Gabriel Fernandez <gabriel.fernandez@foss.st.com>
Change-Id: I4fc724f9c6b7ebd5afaf5d49a554af4cae3c40f4
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/427405
Reviewed-by: Ludovic BARRE <ludovic.barre@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>