platform: stm32mp2: update linker script for MP21 BL2
To ease BL2 memory mapping, that used a reduce memory size, the BL2
linker script is updated. On MP21 the data area will be placed directly
in the RAM section, without a recopy. For that, a new flag is created
(RELOCATE_DATA) and is enabled for MP23 and MP25. The RETRAM section
is renamed to CODE_RODATA, as RETRAM will no more be accurate on MP21
as the RAM section will be spread on RETRAM and SRAM1.
The section tfm_bl2_shared_data is removed as it is not at the correct
place, and its address is retrieved from DT.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I788b0caa599759dc5f45591b87c0215ca1c5ae97
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/446177
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Ludovic BARRE <ludovic.barre@foss.st.com>