platform: stm33mp2: reset: fix cortex a35 reset
For low power mode and for D1 DStandby exit, the ROM code execute a reset
of the cortex-A35 to switch to ARCH64 mode without cortex M33 interaction
so it is not allowed to force the BOOT_CPU1 bit of RCC_CPUBOOTCR.

More over it is more save to send the SW interrupt after the reset
request to be sure the reset is take in account in all the case:
the SW interrupt is used to wake-up the cA35 when VDDCPU is OFF,
when cA35 can't be prefromed. That avoids a possible issue if SW
interruption is cleared by CA35 FW just before to go in WFI and before
reset request.

Change-Id: I55c8d2f31bc0001d5f8bf488af59b05868435250
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/444572
Reviewed-by: Ludovic BARRE <ludovic.barre@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>