Misc: add .checkpatch.conf
For STM32MP2 folder, we use Linux coding style. To check that we need
checkpatch.pl tool, but we need some adaptations for TF-M. That is done
with a file .checkpatch.conf.
This is a copy of the one in TF-A.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I88a763c8e83f9f6d736d6cb8a31348acc70e78d7
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/472953
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>