platform: stm32mp2: refactor firmware public keys hashes management
Rename the OTP words node name so that it reflects what it actually is
in the OTP mapping.

In case MCUBOOT_IMAGE_NUMBER == 3, meaning that MCUBOOT needs to
authenticate and load the Cortex A35 firmware in DDR, the image index
of the DDR firmware to authenticate changes. This firmware has to be
the first authenticated and loaded. Re-route the ID to the correct OTP
field.

Change-Id: I5d97796ce6f1e7cc499321ef4e3f72a46ef7bb77
Signed-off-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/470459
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>