platform: stm32mp215f_dk: allow out-of-tree board directory
Use TARGET_PATH to define STM_DIR if it exists. This is the case for
TF-M secure, but not for non-secure.
For non-secure case stick to the old definition as the board directory
will be copied in the build folder in the stm copied folder there.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I550ef00f285668dc265e8082b52fbb417735ebb5
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/477312
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Ludovic BARRE <ludovic.barre@foss.st.com>