platform: stm32mp235f_dk: allow out-of-tree board directory
Use TARGET_PATH to define STM_DIR if it exists. This is the case for
TF-M secure, but not for non-secure.
For non-secure case stick to the old definition as the board directory
will be copied in the build folder in the stm copied folder there.

Signed-off-by: Yann Gautier <yann.gautier@foss.st.com>
Change-Id: I51b3c83967d36eb9012318343aa79023ade31316
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/477313
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Ludovic BARRE <ludovic.barre@foss.st.com>