platform: stm32mp2: pm: CRC: Reduce CRCized memory size
The total size of the CRC is currently based on the size of the binary
(the lowpower firmware .elf is flatten into a .bin) and linked as data.
We have to set the CRC for the firmware code only.  So the computation
of the size as to be reworked.  For now we forced it one block (32K of
the RETRAM is taken into account for the CRC computation)

Signed-off-by: Simeon Marijon <simeon.marijon@st.com>
Signed-off-by: Raphael Gallais-Pou <raphael.gallais-pou@foss.st.com>
Change-Id: I759e56f3be3108625a478445096d9feaa53f5ab1
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/495564
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Tested-by: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Domain-Review: Patrick DELAUNAY <patrick.delaunay@foss.st.com>
Reviewed-by: Ludovic BARRE <ludovic.barre@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>