platform: stm32mp2: add ramcfg driver
Create a separate RAMCFG driver to enable CRC and signature check
only during Standby (it is needed to execute firmware during Standby
exit) to avoid LP FW execution on simple cortex-M33 reset.

The driver initialize ECC on internal memory BKPSRAM and RETRAM
at init when it is not already done.

Change-Id: I646639ff0c9ea3144fba24f2b77779bb968e77d2
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/499045
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>