platform: stm32mp2: pmic: update support of PWRCTRL_RS
The properties "st,pwrctrl-reset" in device tree enables the independent
reset source for the regulator: when selected PWRCTRL source is active,
regulator is disabled and associated control registers are reset to
default value.

For OpenSTlinux, this feature is used to perform a power cycle on external
boot device when it is needed for ROM code execution, on SD-Card for
example.

With PWRCTRL3=NRSTC1MS + PWR_CPU_ON, the default configuration in the
AN5727, this powerctrl produces a pulse during exit of Stop2, Standby or
Run2/D1 DStandby.

This power cycle for D1 DStandby exit is not required as ROM code is not
executed and can cause functional issue: for SD-Card the supply for IO
domain changes without updating the IO configuration (IOCOMP for example).

To avoid this synchronization issue between IO supply (modified by PMIC)
and IO configuration (as TF-M is not suspended for D1 DStandby), the
PWRCTRL_RS feature is only activated when the STM32MP2 device requests
Stop2 or Standby low power mode, so when stpmic2_reg_pm_suspend() is
called.

Change-Id: I24b4e88e96dedeb4d67a626938e45487d866eaa2
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/504624
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>