platform: stm32mp2: modify clock handling in sk_cipher driver
Bootrom may use security peripherals such as SAES and can disable clocks
outside the scope of the TF-M clock framework. This change modifies the
management of SAES clocks to make it more dynamic. Now, initialization
enables the clocks, and the clocks are released when the peripheral
reset function is called. Consequently, it is now necessary to call the
reset API to disable the clock once the usage is complete.

Signed-off-by: Maxime Méré <maxime.mere@foss.st.com>
Change-Id: I87a9819303ed4dc39d8f1f820b5d39c575f74001
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/512624
Tested-by: Yann GAUTIER <yann.gautier@foss.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>
Domain-Review: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>