platform: stm32mp2: pmic: fix stpmic2_reg_get_voltage for GPO
For regulators which do not support VOUT in MAIN_CR (volt_mask = 0)
or without BYPASS, the register read result is not needed and
stpmic2_reg_get_voltage() always returns the linear result for 0.

Moreover the addresses of volt_cr are not initialized for GPOx,
so the register is read on the address 0x0.

This patch skips this unneeded I2C read and forces 0 as VOUT
default value for the next linear search.

Change-Id: I8a118fe21b91503b67e12ca4037a97f1e83884f5
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/TF-M/trusted-firmware-m/+/515792
ACI: CITOOLS <MDG-smet-aci-reviews@list.st.com>
Reviewed-by: Yann GAUTIER <yann.gautier@foss.st.com>